EveServer/requirements.txt
alembic==0.9.2
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
click==6.7
diamond-mysqlorm==0.1
enum34==1.1.6
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gevent==1.2.0
greenlet==0.4.12
gunicorn==19.7.1
html5lib==0.999999999
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.7.3
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
py==1.4.34
pycparser==2.18
PyMySQL==0.7.10
pyparsing==2.2.0
pytest==3.1.3
pytest-html==1.10.1
python-dateutil==2.6.0
python-editor==1.0.3
redis==2.10.5
requests==2.18.1
selenium==3.0.2
six==1.10.0
SQLAlchemy==1.1.10
urllib3==1.21.1
webencodings==0.5.1
Werkzeug==0.12.2
